logWatcher/requirements.txt
requests
elasticsearch>=7.0.0,<8.0.0
python-dateutil
